% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pedinfo.R
\name{check.pedinfo}
\alias{check.pedinfo}
\title{Check pedigree information.}
\usage{
check.pedinfo(pedinfo)
}
\arguments{
\item{pedinfo}{dataframe.}
}
\description{
\code{check.pedinfo} checks that the pedigree information provided is consistent.
}
\details{
Member ID must be unique. Parents must precede offsprings. Sex information must match parental status, and are coded 1 and 2 for male and female respectively. An error message will be produced only if inconsistencies are found.
}
