% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ibd.R
\name{fgl2relatedness}
\alias{fgl2relatedness}
\title{Score pairwise relatedness.}
\usage{
fgl2relatedness(fgl1p, fgl1m, fgl2p, fgl2m)
}
\arguments{
\item{fgl1p, fgl1m, fgl2p, fgl2m}{positive integer, represents founder genome label.}
}
\value{
A value in [0, 0.5, 1, 2] representing local relatedness coefficient.
}
\description{
\code{fgl2relatedness} determines pairwise relatedness given the four founder genome labels of two individuals at a marker.
}
\examples{
fgl2relatedness(1, 1, 1, 1)
fgl2relatedness(1, 2, 1, 2)
fgl2relatedness(1, 2, 1, 3)
fgl2relatedness(3, 4, 5, 6)
fgl2relatedness(4, 5, 4, 4)
}
